split SequencesDataAPI into separate file sequence_data.py
